baidu-asr-ws/requirements.txt
websocket-client
pyzmq
pyaudio
python-decouple